Integrate tri-solenoid, IoT Hub, Phase 2 contracts, Alchemy RPC mesh,
and mine-with-us docs. Resolve conflicts: unified IoT routes (hub + registry),
full server router mount (nexus/shadow/rpc + tv/single-pane/mining),
combined Vault secret buckets.

+++ contracts/agent-nft/MultiSplitLeasing.sol
@@ -20,7 +20,7 @@ contract MultiSplitLeasing is Ownable, ReentrancyGuard {
     IAgentLeaseReader public agentNFT;
 
   address public treasury;
-  uint8 public treasurySplitBps = 1000; // 10% platform fee on lease revenue
+  uint16 public treasurySplitBps = 1000; // 10% platform fee on lease revenue
 
   event RevenueDistributed(
     uint256 indexed tokenId,
